modify public directory modify index.html change icon to logo.png delete favicon.ico logo192.png and logo512.png create assets folder in src directory move normalize.css and logo.png files delete index.css file
