+++ src/index.js
@@ -1,8 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import { BrowserRouter as Router } from "react-router-dom";
-import "./index.css";
-import "./normalize.css";
+import "./assets/normalize.css";
 import App from "./App";
 import { firebaseConfig } from "./firebase.config";
 import { initializeApp } from "firebase/app";
